блюпринты, рефакторинг
