--- webapp/templates/index.html
@@ -1,8 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <h1>Hi, {{ current_user.username }}!</h1>
-    {% for post in posts %}
-    <div><p>{{ post.owner.username }} says: <b>{{ post.body }}</b></p></div>
-    {% endfor %}
-{% endblock %}
