+++ webapp/templates/user/login.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block content %}
-    <h1>Sign In</h1>
+    <h1>{{ title }}</h1>
     <form action="" method="post" novalidate>
         {{ form.hidden_tag() }}
         <div class="form-group">
@@ -23,6 +23,6 @@
             {{ form.remember_me.label(class_='form-check-label') }}
         </div>
         <p>{{ form.submit() }}</p>
-        <p>New User? <a href="{{ url_for('register') }}">Click to Register!</a></p>
+        <p>New User? <a href="{{ url_for('user.register') }}">Click to Register!</a></p>
     </form>
 {% endblock %}
