+++ webapp/templates/user/profile.html
@@ -5,9 +5,6 @@
         <tr valign="top">
             <td>
                 <h1>User: {{ user.username }}</h1>
-                {% if user.about_me %}
-                    <p>{{ user.about_me }}</p>
-                {% endif %}
                 {% if user.last_seen %}
                     <p>Last seen on: {{ user.last_seen }}</p>
                 {% endif %}
