+++ webapp/templates/user/register.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block content %}
-    <h1>Register</h1>
+    <h1>{{ title }}</h1>
     <form action="" method="post">
         {{ form.hidden_tag() }}
         <p>
